+++ server/src/main/java/com/blocklang/release/controller/ReleaseController.java
@@ -23,10 +23,7 @@ import com.blocklang.exception.ResourceNotFoundException;
 import com.blocklang.release.constant.ReleaseResult;
 import com.blocklang.release.data.NewReleaseTaskParam;
 import com.blocklang.release.model.ProjectReleaseTask;
-import com.blocklang.release.model.ProjectTag;
 import com.blocklang.release.service.BuildService;
-import com.blocklang.release.service.GitToolService;
-import com.blocklang.release.service.ProjectBuildService;
 import com.blocklang.release.service.ProjectReleaseTaskService;
 import com.blocklang.release.service.ProjectTagService;