+++ server/src/main/java/com/blocklang/release/service/BuildService.java
@@ -5,9 +5,7 @@ import com.blocklang.release.model.ProjectReleaseTask;
 import com.blocklang.release.task.AppBuildContext;
 
 /**
- * 简单封装 npm install，maven install 和 dojo build 等 build 相关操作
- * 
- * TODO: 将这些单个的服务最终合并成一个 build 服务？
+ * 实际构建服务
  * 
  * @author Zhengwei Jin
  *
@@ -20,13 +18,13 @@ public interface BuildService {
 	 * @param appBuildContext
 	 * @return 执行成功返回 <code>true</code>，执行失败返回<code>false</code>
 	 */
-	boolean runNpmInstall(AppBuildContext appBuildContext);
-
-	boolean runDojoBuild(AppBuildContext appBuildContext);
-
-	boolean copyDojoDistToSpringBoot(AppBuildContext appBuildContext);
-
-	boolean runMavenInstall(AppBuildContext appBuildContext);
+//	boolean runNpmInstall(AppBuildContext appBuildContext);
+//
+//	boolean runDojoBuild(AppBuildContext appBuildContext);
+//
+//	boolean copyDojoDistToSpringBoot(AppBuildContext appBuildContext);
+//
+//	boolean runMavenInstall(AppBuildContext appBuildContext);
 	
 	/**
 	 * 构建 block lang 项目，该方法属于异步方法。