+++ server/src/main/java/com/blocklang/release/service/impl/BuildServiceImpl.java
@@ -1,0 +1,18 @@
+package com.blocklang.release.service.impl;
+
+import org.springframework.stereotype.Service;
+
+import com.blocklang.develop.model.Project;
+import com.blocklang.release.model.ProjectReleaseTask;
+import com.blocklang.release.service.BuildService;
+
+@Service
+public class BuildServiceImpl implements BuildService {
+
+	@Override
+	public void build(Project project, ProjectReleaseTask releaseTask) {
+		// TODO Auto-generated method stub
+
+	}
+
+}