+++ server/src/main/java/com/blocklang/release/service/impl/ProjectReleaseTaskServiceImpl.java
@@ -1,0 +1,17 @@
+package com.blocklang.release.service.impl;
+
+import org.springframework.stereotype.Service;
+
+import com.blocklang.release.model.ProjectReleaseTask;
+import com.blocklang.release.service.ProjectReleaseTaskService;
+
+@Service
+public class ProjectReleaseTaskServiceImpl implements ProjectReleaseTaskService {
+
+	@Override
+	public ProjectReleaseTask save(ProjectReleaseTask projectReleaseTask) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+}